Feat: jwt 내 swagger 엔드포인트 허용
- Entity 속성 수정 전 마지막 커밋

+++ battle-people-server/src/main/java/com/woowahanrabbits/battle_people/domain/balancegame/service/BalanceGameService.java
@@ -9,7 +9,7 @@ import com.woowahanrabbits.battle_people.domain.user.domain.User;
 import com.woowahanrabbits.battle_people.domain.vote.domain.UserVoteOpinion;
 
 public interface BalanceGameService {
-	void addBalanceGame(CreateBalanceGameRequest createBalanceGameRequest, int userId);
+	void addBalanceGame(CreateBalanceGameRequest createBalanceGameRequest, Long userId);
 
 	List<BalanceGameResponse> getBalanceGameByConditions(Integer category, int status, int page, User user);
